fix(contracts): disambiguate TFHE.asEuint64(0) and tighten helpers

+++ contracts/contracts/PrivateLendingPool.sol
@@ -118,7 +118,7 @@ contract PrivateLendingPool {
      */
     function _accrue(address user) private {
         euint64 oldDebt = debts[user];
-        ebool hasDebt = TFHE.gt(oldDebt, TFHE.asEuint64(0));
+        ebool hasDebt = TFHE.gt(oldDebt, TFHE.asEuint64(uint64(0)));
         euint64 onePlusRate = TFHE.add(ratePerBlockRay, MathEncrypted.RAY());
         euint64 tmp = MathEncrypted.scaleMul(oldDebt, onePlusRate);
         euint64 newDebt = MathEncrypted.scaleDiv(tmp, MathEncrypted.RAY());
